+++ src/jest-snapshot-serializer.ts
@@ -1,0 +1,13 @@
+import {isOpt, toString} from './Opt';
+
+interface JestSnapshotSerializer {
+  serialize(val: unknown): string;
+  test(val: unknown): boolean;
+}
+
+const jestSnapshotSerializer: JestSnapshotSerializer = {
+  serialize: toString,
+  test: isOpt,
+};
+
+export default jestSnapshotSerializer;